hack to get storage connection info swapped out for test db config
Trying to get the WithCurrentUser trait working. By simply enabling it, it makes
the ->connection call too late. Calling ->schema to get an object to call ->connection
on sends it in to the "per request" stuff and that schema already has catalyst config
connection info, so $c has a schema connected to the wrong db.

This hacks in to the PerRequest stuff to get the desired connection info in to the
schema before it gets cloned